move task in celery app
